v. 2.17.9
- kernel replaced with core
- Added client verification to telemetry
- Fixed currency loading
- Methods for working with cooldown are moved to a separate class
- Removed reply to message when returning from afk
- Added counting of users and balance in the user database
- Added updating of the twitch client token when disconnecting
- Fixed delay in CpuUsage on Windows
- Fixed counting of new users
- Improved message sending system
- Fixed saving of buffer and rate
- Fixed loading of rate from file

+++ Bot/Core/Commands/List/AI.cs
@@ -39,8 +39,8 @@ namespace butterBror.Core.Commands.List
             {
                 if (MessageProcessor.GetArgument(data.Arguments, "chat") is null)
                 {
-                    float currency = butterBror.Bot.BankDollars / butterBror.Bot.Coins;
-                    float cost = 0.5f / currency;
+                    decimal currency = butterBror.Bot.BankDollars / butterBror.Bot.Coins;
+                    decimal cost = 0.5m / currency;
 
                     int coins = -(int)cost;
                     int subcoins = -(int)((cost - coins) * 100);
